pyproject.toml
[project]
name = "gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.35",
    "crypto>=1.4.1",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.117.1",
    "hf-xet>=1.1.10",
    "pyjwt[crypto]>=2.10.1",
    "pymupdf>=1.26.4",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
]
